% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_clean}
\alias{simulate_clean}
\title{Run Clean Completion Time Simulation}
\usage{
simulate_clean(n = 250, reps = 1000, rate = 1)
}
\arguments{
\item{n}{Sample size.}

\item{reps}{Number of simulation replications.}

\item{rate}{Rate parameter for the exponential distribution.}
}
\value{
A data frame comparing estimator performance.
}
\description{
Simulates clean completion times from an exponential distribution.
}
