% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_contaminated}
\alias{simulate_contaminated}
\title{Run Contaminated Completion Time Simulation}
\usage{
simulate_contaminated(n = 250, reps = 1000, contamination_prop = 0.1)
}
\arguments{
\item{n}{Sample size.}

\item{reps}{Number of simulation replications.}

\item{contamination_prop}{Proportion of contaminated observations.}
}
\value{
A data frame comparing estimator performance.
}
\description{
Simulates completion times with 90% clean observations and 10% contaminated observations.
}
